+++ sdk/resourcemanager/elasticsan/armelasticsan/CHANGELOG.md
@@ -1,5 +1,15 @@
 # Release History
 
+## 1.1.0-beta.1 (2024-04-22)
+### Features Added
+
+- New enum type `PolicyEnforcement` with values `PolicyEnforcementDisabled`, `PolicyEnforcementEnabled`
+- New struct `AutoScaleProperties`
+- New struct `ScaleUpProperties`
+- New field `AutoScaleProperties` in struct `Properties`
+- New field `AutoScaleProperties` in struct `UpdateProperties`
+
+
 ## 1.0.0 (2024-01-26)
 ### Other Changes